Don't rely on BlankNode.ntriplesString() making sense.
As pointed out in #42, the ntriplesString() does not need to match
the internal identifier. 

The tests now also assume that RDFNodes in a Triple are equal to those
passed into the RDFTermFactory.createTriple() - as those
terms were also made with the same RDFTermFactory and the Triple
is not (yet) inserted into a graph.