COMMONSRDF-6 Simplify internalIdentifier salts
BlankNodeImpl constructor take a UUID as salt, which is created for each
SimpleRDFTermFactory instance.  SimpleRDFTermFactory keeps a random
UUIDv4 per instance for this purpose.

(This may not be effective if loads of factories are made, but this is
the simple implementation which is not required to be pre-emptively
efficient.)

Constructor BlankNodeImpl() without arguments still uses an
internal AtomicLong, now combined with an internal UUID salt.
This could in theory just do

    this.id = UUID.randomUUID().toString()

and not keep any counter, but it is much more conceivable that say
10.000 BlankNodeImpls will be made, than 10.000 SimpleRDFTermFactories.
I did not do any performance testing, but kept the AtomicLong mechanism
as I didn't feel a need to change it now.