Source code of examples from User guide
See
http://mail-archives.apache.org/mod_mbox/commonsrdf-dev/201504.mbox/%3CCAOfJQJ0ifthZMeFna6jMPCcZGdaGaKVwc2djuJMi02Vz1odPvg%40mail.gmail.com%3E

NOT part of the build as it should not be released
to Maven Central.  Also non-Apache groupId, package etc.
as it is just an example.

TODO: Some kind of way to include the <module> in the <profile>
when it is NOT the -Papache-release profile? Would be great
if the <version> were automatically maintained or at least
the test regularly run in development build.