Added SimpleRDFTerm marker interface
Now internallyMap() will fail early on mapping
any RDFTerm objects which are not IRI, BlankNode or Literal
(before they would have been added 'as-is')

This also fixes a bug in that Graph would have remapped
any of the Types instances to an IRIImpl.

I added SimpleRDFTerm interface as a package-protected
interface within SimpleRDFTermFactory - thus it should not
be easy to create any additional SimpleRDFTerm subclasses.