Create our own BlankNode for testing
The test was improved to make the g1 and g2 with its own BlankNode instances - each graph then having _:b1 and _:b2 - but with different uniqueReference().

This would also test that multiple graph.add(s,p,o) works with foreign BlankNode instances, but without taking a stance on COMMONSRDF-15 - we don't care how it does the mapping (if any) - and don't even require that they have the same uniqueReference() after adding.

Iterating g1 and g2 to add to g3 should not assume those BlankNodes are the same based on their triplesString() - but differ by their uniqueReference()