TripleOrQuad -> TripleLike, GraphLike
TripleLike is a generalized Triple (e.g. as seen in Jena), supporting
any RDFTerm in any of the subject/predicate/object positions.
This can be useful for parsers, and also serves as a common
super-type for Quad and Triple.

Equivalently, QuadLike extends TripleLike as a generalized Quad, supporting
any RDFTerm as its graphName.